+++ Frends.MicrosoftSQL.BulkInsert/Frends.MicrosoftSQL.BulkInsert.Tests/UnitTests.cs
@@ -420,6 +420,64 @@ public class UnitTests
         }
     }
 
+    [TestMethod]
+    public async Task TestBulkInsert_ColumnMapping()
+    {
+        var columnMappings = new List<ColumnMapping>()
+        {
+            ColumnMapping.JsonPropertyOrder,
+            ColumnMapping.JsonPropertyNames,
+            ColumnMapping.ManualColumnMapping
+        };
+
+        var options = new Options()
+        {
+            SqlTransactionIsolationLevel = SqlTransactionIsolationLevel.ReadCommitted,
+            CommandTimeoutSeconds = 60,
+            FireTriggers = false,
+            KeepIdentity = true,
+            NotifyAfter = 0,
+            ConvertEmptyPropertyValuesToNull = true,
+            KeepNulls = false,
+            TableLock = false,
+        };
+
+        var json = @"[
+                  {
+                    ""Id"": 1,
+                    ""FirstName"": ""Etu"",
+                    ""LastName"": ""Suku""
+                  },
+                  {
+                    ""Id"": 2,
+                    ""LastName"": ""Suku"",
+                    ""FirstName"": ""Etu""
+                  },
+                  {
+                    ""FirstName"": ""First"",
+                    ""LastName"": ""Last"",
+                    ""Id"": 3
+                  }
+                ]";
+
+        var manualMapping = @"{ ""Id"":""Id"", ""FirstName"": ""FirstName"", ""LastName"": ""LastName"" }";
+
+        foreach (var columnMapping in columnMappings)
+        {
+            var input = new Input
+            {
+                ConnectionString = _connString,
+                TableName = _tableName,
+                InputData = json,
+                ColumnMapping = columnMapping,
+                ManualColumnMapping = manualMapping
+            };
+
+            var result = await MicrosoftSQL.BulkInsert(input, options, default);
+            Assert.IsTrue(result.Success);
+        }
+    }
+
     private static int GetRowCount()
     {
         using var connection = new SqlConnection(_connString);