+++ Frends.MicrosoftSQL.BulkInsert/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [3.0.0] - 2025-01-15
+### Added
+- [Breaking] Added parameters ColumnMapping and ManualColumnMapping.
+- Added column mapping feature which allows user to select from JsonPropertyNames, JsonPropertyOrder and ManualColumnMapping options how the column mapping is handled in bulk insert. Default behavior will remain the same JsonPropertyOrder.
+
 ## [2.2.0] - 2024-09-10
 ### Changed
 - Updated Options.NotifyAfter property to be set dynamically based on the total row count, with a minimum value of 1, ensuring rowsCopied is updated correctly.